+++ tests/testthat/test_tabpfn_classif_tabpfn.R
@@ -1,3 +1,9 @@
+test_that("Python config is valid", {
+  cnfg = reticulate::py_config()
+  env = basename(cnfg$pythonhome)
+  expect_equal(env, "r-reticulate")
+})
+
 test_that("autotest", {
   learner = lrn("classif.tabpfn")
   expect_learner(learner)