+++ NEWS.md
@@ -8,6 +8,7 @@
 * feat: Added learner `LearnerClassifAdaBoosting`.
 * feat: Added learners `LearnerClassifEvtree` and `LearnerRegrEvtree`.
 * feat: Added learner `LearnerClassifKnn`.
+* feat: Added learner `LearnerClassifFastai`.
 
 # mlr3extralearners 1.1.0